update types, relations, and indecies for the db

+++ types/user.go
@@ -36,7 +36,7 @@ type UpdateUserInput struct {
 	VerificationToken   string    `json:"-"`
 	Verified            bool      `json:"-"`
 	VerifiedAt          time.Time `json:"-"`
-	SigninCount         int64     `json:"-"`
+	SigninCount         uint      `json:"-"`
 	CurrentSigninAt     time.Time `json:"-"`
 	CurrentSigninIP     string    `json:"-"`
 	LastSigninAt        time.Time `json:"-"`
